RDM-5903: Implement and consume external GET event details for case endpoint in API v2
