+++ src/main/java/uk/gov/hmcts/ccd/v2/external/resource/CaseEventsResource.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.ccd.v2.internal.resource;
+package uk.gov.hmcts.ccd.v2.external.resource;
 
 import java.util.List;
 import lombok.Data;
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 import lombok.NonNull;
 import org.springframework.hateoas.RepresentationModel;
 import uk.gov.hmcts.ccd.domain.model.std.AuditEvent;
-import uk.gov.hmcts.ccd.v2.internal.controller.UICaseController;
+import uk.gov.hmcts.ccd.v2.external.controller.CaseController;
 
 import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
 import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
@@ -15,12 +15,12 @@ import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
 @Data
 @EqualsAndHashCode(callSuper = true)
 @NoArgsConstructor
-public class UICaseEventsResource extends RepresentationModel {
+public class CaseEventsResource extends RepresentationModel {
 
     private List<AuditEvent> auditEvents;
 
-    public UICaseEventsResource(@NonNull String caseReference, @NonNull List<AuditEvent> listOfAuditEvents) {
+    public CaseEventsResource(@NonNull String caseReference, @NonNull List<AuditEvent> listOfAuditEvents) {
         this.auditEvents = listOfAuditEvents;
-        add(linkTo(methodOn(UICaseController.class).getCaseEvents(caseReference)).withSelfRel());
+        add(linkTo(methodOn(CaseController.class).getCaseEvents(caseReference)).withSelfRel());
     }
 }
